UPDATE - Remove redundant code
