+++ src/Pages/CreateRoom/components/CreateBEQuestions.jsx
@@ -3,130 +3,14 @@ import React, { useRef, useState } from 'react';
 import ButtonStyled from '../../../components/Button';
 import localFilesAPI from '../../../utils/api/localFilesAPI';
 import * as St from '../styles';
+import MaxSumbitTimes from './MaxSumbitTimes';
 import Testcase from './Testcase';
+import UploadImage from './UploadImage';
+import { addBEQuestion, addTestcase, editInputTestcase, editOutputTestcase } from './helper.js';
 
 const CreateBEQuestions = ({ questions, setQuestions }) => {
     const inputRef = useRef(null);
     const [imageUrls, setImageUrls] = useState([]);
-
-    // Functions
-    const createQuestion = () => {
-        setQuestions((prev) => [
-            ...prev,
-            {
-                maxSubmitTimes: 0,
-                questionImage: '',
-                testcases: [{ input: '', output: '' }],
-            },
-        ]);
-    };
-
-    const createTestcase = (questionIdx) => {
-        setQuestions((prev) => {
-            const { testcases, ...rest } = prev[questionIdx]; // destructure testcases array and the rest of the question object
-            const newTestcases = [...testcases, { input: '', output: '' }]; // add a new testcase to the testcases array
-            const updatedQuestion = { ...rest, testcases: newTestcases }; // create a new object with updated testcases array
-            const updatedQuestions = [
-                ...prev.slice(0, questionIdx),
-                updatedQuestion,
-                ...prev.slice(questionIdx + 1),
-            ]; // replace the old question object with the updated one
-            return updatedQuestions;
-        });
-    };
-
-    const editInputTestcase = (value, questionIdx, testcaseIdx) => {
-        setQuestions((prev) => {
-            let copy = [...prev];
-            copy[questionIdx].testcases[testcaseIdx].input = value;
-            return copy;
-        });
-    };
-    const editOutputTestcase = (value, questionIdx, testcaseIdx) => {
-        setQuestions((prev) => {
-            let copy = [...prev];
-            copy[questionIdx].testcases[testcaseIdx].output = value;
-            return copy;
-        });
-    };
-
-    const handleIncrease = (questionIdx) => {
-        setQuestions((prev) => {
-            let copy = [...prev];
-            copy[questionIdx].maxSubmitTimes += 1;
-            return copy;
-        });
-    };
-    const handleDecrease = (questionIdx) => {
-        setQuestions((prev) => {
-            let copy = [...prev];
-            copy[questionIdx].maxSubmitTimes -= 1;
-            return copy;
-        });
-    };
-
-    const uploadImages = async (files, questionIdx) => {
-        try {
-            const file = files[0];
-            const res = await localFilesAPI.uploadQuestionFile(file);
-            const reader = new FileReader();
-
-            reader.onload = (event) => {
-                const imageUrl = event.target.result;
-                setImageUrls((prev) => {
-                    let copy = [...prev];
-                    copy[questionIdx] = imageUrl;
-                    return copy;
-                });
-
-                setQuestions((prev) => {
-                    let copy = [...prev];
-                    copy[questionIdx].questionImage = res.data.data[0].id;
-                    return copy;
-                });
-            };
-
-            reader.readAsDataURL(file);
-        } catch (error) {
-            console.log('Error: ', error);
-        }
-    };
-
-    // Components
-    const MaxSumbitTimes = ({ question, questionIdx }) => {
-        return (
-            <div>
-                <label htmlFor="maxSubmitTimes">Maximum Submit Time:</label>
-                <St.NumberInput>
-                    <div onClick={() => handleDecrease(questionIdx)}>-</div>
-                    <p>{question.maxSubmitTimes}</p>
-                    <div onClick={() => handleIncrease(questionIdx)}>+</div>
-                </St.NumberInput>
-            </div>
-        );
-    };
-
-    const UploadImages = ({ question, questionIdx }) => {
-        return (
-            <div className="my-4">
-                <label>Image</label>
-                <input
-                    id="questionImage"
-                    type="file"
-                    hidden
-                    ref={inputRef}
-                    onChange={(e) => uploadImages(e.target.files, questionIdx)}
-                />
-                {question.questionImage ? (
-                    <St.PreviewImage src={imageUrls[questionIdx]} alt="Image" />
-                ) : (
-                    <St.UploadImage onClick={() => inputRef.current.click()}>
-                        + <span>Upload your image here...</span>
-                    </St.UploadImage>
-                )}
-            </div>
-        );
-    };
 
     return (
         <>
@@ -136,8 +20,18 @@ const CreateBEQuestions = ({ questions, setQuestions }) => {
 
                     <div className="row">
                         <div className="col-md-6">
-                            <MaxSumbitTimes question={question} questionIdx={questionIdx} />
-                            <UploadImages question={question} questionIdx={questionIdx} />
+                            <MaxSumbitTimes
+                                question={question}
+                                questionIdx={questionIdx}
+                                setQuestions={setQuestions}
+                            />
+                            <UploadImage
+                                question={question}
+                                questionIdx={questionIdx}
+                                setQuestions={setQuestions}
+                                setImageUrls={setImageUrls}
+                                imageUrls={imageUrls}
+                            />
                         </div>
 
                         <div className="col-md-6">
@@ -149,12 +43,13 @@ const CreateBEQuestions = ({ questions, setQuestions }) => {
                                     questionIdx={questionIdx}
                                     handleInput={editInputTestcase}
                                     handleOutput={editOutputTestcase}
+                                    setQuestions={setQuestions}
                                 />
                             ))}
                             <div className="d-grid gap-2 ">
                                 <ButtonStyled
                                     buttonType="outline"
-                                    onClick={() => createTestcase(questionIdx)}
+                                    onClick={() => addTestcase(questionIdx, setQuestions)}
                                 >
                                     + Add testcase
                                 </ButtonStyled>
@@ -166,7 +61,7 @@ const CreateBEQuestions = ({ questions, setQuestions }) => {
             ))}
 
             <div className="d-grid gap-2 my-2">
-                <ButtonStyled buttonType="dashed" onClick={createQuestion}>
+                <ButtonStyled buttonType="dashed" onClick={() => addBEQuestion(setQuestions)}>
                     + Add questions
                 </ButtonStyled>
             </div>
